moved applications import from lib_old
